Enroll to subject is working. TODO: enroll to exam needs crash fix.

+++ src/main/java/com/naveen/learningnavigator/dto/AddSubjectDto.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class AddSubjectDto {
     @NotNull
-    private int subjectId;
+    private Long subjectId;
     @NotEmpty
     private String subjectName;
 }
